requirements.txt
django==5.1
djangorestframework==3.15.2
google-api-python-client==2.142